Eliminate selection interface as there is only one implementor.
