+++ src/main/java/com/esgdev/sparkpaint/engine/selection/Selection.java
@@ -1,32 +1,171 @@
 package com.esgdev.sparkpaint.engine.selection;
 
 import java.awt.*;
+import java.awt.geom.AffineTransform;
+import java.awt.geom.GeneralPath;
 import java.awt.image.BufferedImage;
 
-public interface Selection {
-    BufferedImage getContent();
 
-    void setContent(BufferedImage content);
+/// Represents the selection area and its content in the drawing canvas.
+public class Selection {
+    private GeneralPath path;
+    private BufferedImage content;
+    private boolean transparent;
 
-    boolean isTransparent();
+    public Selection(Rectangle rect, BufferedImage content) {
+        this.path = new GeneralPath();
+        path.append(rect, false);
+        this.content = content;
+    }
 
-    void setTransparent(boolean transparent);
+    public Selection(GeneralPath path, BufferedImage content) {
+        this.path = path;
+        this.content = content;
+    }
 
-    boolean contains(Point point);
+    public GeneralPath getPath() {
+        return path;
+    }
 
-    void clear();
+    public void setPath(GeneralPath path) {
+        this.path = path;
+    }
 
-    void clearOutline();
+    public BufferedImage getContent() {
+        return content;
+    }
 
-    boolean hasOutline();
+    public void setContent(BufferedImage content) {
+        this.content = content;
+    }
 
-    void rotate(int degrees);
+    public boolean isTransparent() {
+        return this.transparent;
+    }
 
-    Rectangle getBounds();
+    public void setTransparent(boolean transparent) {
+        this.transparent = transparent;
+    }
 
-    void delete(Graphics2D g2d, Color canvasBackground);
+    public boolean contains(Point point) {
+        return path != null && path.contains(point);
+    }
 
-    void drawSelectionContent(Graphics2D g2d, double zoomFactor);
+    public void clear() {
+        path = null;
+        content = null;
+    }
 
-    void drawSelectionOutline(Graphics2D g2d, double zoomFactor);
+
+    public void clearOutline() {
+        path = null;
+    }
+
+
+    public boolean hasOutline() {
+        return path != null;
+    }
+
+
+    public void rotate(int degrees) {
+        if (content == null) return;
+
+        BufferedImage original = content;
+        int width = original.getWidth();
+        int height = original.getHeight();
+
+        // Create new rotated image
+        BufferedImage rotated = new BufferedImage(
+                degrees % 180 == 0 ? width : height,
+                degrees % 180 == 0 ? height : width,
+                BufferedImage.TYPE_INT_ARGB
+        );
+
+        Graphics2D g2d = rotated.createGraphics();
+        g2d.translate((rotated.getWidth() - width) / 2,
+                (rotated.getHeight() - height) / 2);
+        g2d.rotate(Math.toRadians(degrees), width / 2.0, height / 2.0);
+        g2d.drawImage(original, 0, 0, null);
+        g2d.dispose();
+
+        content = rotated;
+
+        // Rotate the path
+        Rectangle bounds = path.getBounds();
+        AffineTransform transform = new AffineTransform();
+        transform.rotate(Math.toRadians(degrees), bounds.getCenterX(), bounds.getCenterY());
+        path.transform(transform);
+    }
+
+
+    public Rectangle getBounds() {
+        if (path != null) {
+            return path.getBounds();
+        }
+        return null;
+    }
+
+    public void delete(Graphics2D g2d, Color background) {
+        // Use AlphaComposite.CLEAR instead of filling with background color
+        Composite originalComposite = g2d.getComposite();
+        g2d.setComposite(AlphaComposite.getInstance(AlphaComposite.CLEAR));
+
+        // Fill the selection area with transparency
+        GeneralPath path = this.getPath();
+        if (path != null) {
+            g2d.fill(path);
+        } else {
+            g2d.fill(getBounds());
+        }
+
+        g2d.setComposite(originalComposite);
+    }
+
+    public void drawSelectionContent(Graphics2D g2d, double zoomFactor) {
+        if (content != null && path != null) {
+            Rectangle bounds = path.getBounds();
+            g2d.drawImage(content, bounds.x, bounds.y, null);
+        }
+    }
+
+    public void drawSelectionOutline(Graphics2D g2d, double zoomFactor) {
+        if (path == null) {
+            return;
+        }
+
+        float[] dashPattern = {5, 5}; // Dash pattern: 5px dash, 5px gap
+
+        BasicStroke dottedStroke1 = new BasicStroke(
+                1, BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER,
+                10.0f, dashPattern, 0);
+
+        BasicStroke dottedStroke2 = new BasicStroke(
+                1, BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER,
+                10.0f, dashPattern, 5);
+
+        // Scale the path to match the zoom factor
+        GeneralPath scaledPath = new GeneralPath(path);
+        scaledPath.transform(AffineTransform.getScaleInstance(zoomFactor, zoomFactor));
+
+        // Draw the selection outline with the first dash pattern
+        g2d.setColor(Color.BLACK);
+        g2d.setStroke(dottedStroke1);
+        g2d.draw(scaledPath);
+
+        // Draw the selection outline with the second dash pattern
+        g2d.setColor(Color.WHITE);
+        g2d.setStroke(dottedStroke2);
+        g2d.draw(scaledPath);
+    }
+
+    public static Selection createRectangular(Rectangle rect, BufferedImage content) {
+        GeneralPath path = new GeneralPath();
+        path.moveTo(rect.x, rect.y);
+        path.lineTo(rect.x + rect.width, rect.y);
+        path.lineTo(rect.x + rect.width, rect.y + rect.height);
+        path.lineTo(rect.x, rect.y + rect.height);
+        path.closePath();
+
+        return new Selection(path, content);
+    }
 }
